requirements.txt
flask
pillow
fpdf
